requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.6.0
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
black==22.10.0
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cytoolz==0.12.0
dataclassy==0.11.1
Django==3.2.18
django-crispy-forms==2.0
djongo==1.3.6
dnspython==2.3.0
eip712==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.3
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
frozendict==2.3.8
frozenlist==1.3.1
hexbytes==0.2.3
hypothesis==6.27.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lru-dict==1.1.8
moralis==0.1.37
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
optional-django==0.1.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.10.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.19.5
wrapt==1.14.1
yarl==1.8.1
zipp==3.15.0